Add files via upload
conways game of life (AL) exercise
https://rustwasm.github.io/docs/book/game-of-life/implementing.html

"Representing each cell with a byte makes iterating over cells easy, but it comes at the cost of wasting memory. Each byte is eight bits, but we only require a single bit to represent whether each cell is alive or dead. Refactor the data representation so that each cell uses only a single bit of space."
The files with refactored code are uploaded.

Random initial conditions, periodic boundary conditions.

run 'wasm-pack build' to bind it all together

To do this tutorial download get rustc, cargo (comes with rust), cargo-generate, npm and webpack.
Theres one frustrating bug in the tutorial, a single comma (at the bottom of the package.json file in pkg/ which is built by wasm-pack) has to be deleted. It's next to a },  . The compiler complains that the } is extraneous, no, it's the comma.  ,